Converted template functions
